adding projects

+++ assignment/src/main/java/spring/OrderDAO.java
@@ -4,11 +4,12 @@ import java.util.List;
 
 import hibernate.Orders;
 
+
 public interface OrderDAO {
 
 	void save(Orders o);
 	void update(Orders o);
 	void delete(int id);
-	Orders get(long id);
+	Orders get(int id);
 	List<Orders> getAll();
 }
